+++ CHANGES.rst
@@ -2,6 +2,7 @@
 -----
 
 * views.get_next_url now considers subdomains as non-external.
+* sqlalchemy.BaseMixin now provides a new query class with ``one_or_none``
 
 0.4.0
 -----